collapse scope selection into filter panel
